Change lasagna page's name and create template
I changed the lasagna page's name to match the online recipe I'm using
for it.

I added the markup structure for the recipe content in the lasagna page
to use as a template for the other recipe pages to save time when I
write those pages in a later commit.